Extract the namespace fixtures into a reusable trait
