+++ storage/src/test/scala/uk/ac/wellcome/storage/store/StoreFixtures.scala
@@ -2,16 +2,16 @@ package uk.ac.wellcome.storage.store
 
 import org.scalatest.{Assertion, Matchers}
 import uk.ac.wellcome.fixtures.TestWith
+import uk.ac.wellcome.storage.store.fixtures.NamespaceFixtures
 
-trait StoreFixtures[Id, T, Namespace, StoreContext] extends Matchers {
-  type StoreImpl = Store[Id, T]
+trait StoreFixtures[Ident, T, Namespace, StoreContext] extends Matchers with NamespaceFixtures[Ident, Namespace] {
+  type StoreImpl = Store[Ident, T]
 
-  def withStoreImpl[R](storeContext: StoreContext, initialEntries: Map[Id, T])(testWith: TestWith[StoreImpl, R]): R
+  def withStoreImpl[R](storeContext: StoreContext, initialEntries: Map[Ident, T])(testWith: TestWith[StoreImpl, R]): R
 
-  def withNamespace[R](testWith: TestWith[Namespace, R]): R
   def withStoreContext[R](testWith: TestWith[StoreContext, R]): R
 
-  def withStoreImpl[R](initialEntries: Map[Id, T])(testWith: TestWith[StoreImpl, R]): R =
+  def withStoreImpl[R](initialEntries: Map[Ident, T])(testWith: TestWith[StoreImpl, R]): R =
     withStoreContext { storeContext =>
       withStoreImpl(storeContext, initialEntries) { storeImpl =>
         testWith(storeImpl)
@@ -23,7 +23,6 @@ trait StoreFixtures[Id, T, Namespace, StoreContext] extends Matchers {
       testWith(storeImpl)
     }
 
-  def createId(implicit namespace: Namespace): Id
   def createT: T
 
   def assertEqualT(original: T, stored: T): Assertion =
